laravel/lite-app/app/Http/Controllers/RecommendationScripts/requirements.txt
psycopg2-binary
langcodes
pandas
numpy
<<<<<<< HEAD
scikit-learn
=======
pycountry
scikit-learn
>>>>>>> db6aaa4 (fix bugs)
